cambios en los repositorios

+++ src/main/java/com/example/Proyecto/Repositorios/RolRepository.java
@@ -1,9 +1,7 @@
 package com.example.Proyecto.Repositorios;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.Proyecto.Clases.Rol;
 
-public interface RolRepository extends JpaRepository<Rol, Long> {
-
+public interface RolRepository extends JpaRepository<Rol, Integer> {
 }
